Working user authentication
I got a note from a Youtuber that Django uses a special password policy so I tried to apply this aswell at first. Also that it was easier to extend the base.html to the login.html and register.html. But I firstly included the html in both login and register making it unresponsive. So I fixed this