Creating a login-logout route
I wanted to make sure that users can be redirected when they logout. I had a problem with that so I decided to quickly add an action into my settings.py that redirects the users when logging in and when logging out. The styling was just to test how it worked. I'll be changing this later on.