Added views + updated url
I updated urls.py to include routes for deleting (<int:pk>/delete/) articles, linking them to the new views I just added too.

I created edit_article in views.py, allowing authenticated users to update their own articles. This view fetches the article using get_object_or_404, limiting access to the article’s author. It uses ArticleForm and renders edit_article.html. A bit of the same goes for the users that want to delete an article that's linked to their account. Calling it ArticleDeleteView

The html files are the ones that I'll be adding and editing later on. For now I already made the files but they just need to updated so an extention from base.html will be placed there.